refactor: refactored logger, into 3 classes, one to handle the loggin itself, and one to handle options, notifiers, etc...
