+++ examples/basic/basic-example.js
@@ -3,6 +3,8 @@
 
 const loggin = require('../../');
 
+loggin.debug();
+
 // Create a logger with a set of options
 // `level`   - Set output level to DEBUG (all logs will output)
 // `color`   - Set color for all notifiers
