+++ package.json
@@ -58,7 +58,7 @@
     "strif": "^1.1.4"
   },
   "devDependencies": {
-    "@babel/cli": "^7.6.4",
+    "@babel/cli": "^7.13.10",
     "@babel/core": "^7.6.4",
     "@babel/preset-env": "^7.6.3",
     "browserify": "^16.5.0",
